Fix bug in generation arguments
